% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{nirs4allio_version}
\alias{nirs4allio_version}
\title{Report the nirs4allio binding version}
\usage{
nirs4allio_version()
}
\value{
A length-one character vector with the binding version.
}
\description{
Returns the version of the \code{nirs4allio} R binding as a character scalar.
This is the binding's own version and is independent of the underlying Rust
\code{nirs4all-io} core version.
}
\examples{
nirs4allio_version()

}
\seealso{
\code{\link[=nirs4allio_native_available]{nirs4allio_native_available()}}.
}
